@@ -42,7 +42,7 @@
         <beetl>2.8.5</beetl>
         <apollo>1.3.0</apollo>
         <jsms-client>1.2.0</jsms-client>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <php-common.es>2.1.3-SNAPSHOT</php-common.es>
         <php-common.jdbc>2.1.3-SNAPSHOT</php-common.jdbc>
         <php-common.jpa>2.1.3-SNAPSHOT</php-common.jpa>
