@@ -4,8 +4,8 @@ import com.php25.common.core.dto.ResultDto;
 import com.php25.common.core.util.AssertUtil;
 import com.php25.common.core.util.StringUtil;
 import com.php25.mediamicroservice.client.bo.ImgBo;
-import com.php25.mediaservice.client.bo.ImgDto;
 import com.php25.mediamicroservice.client.rpc.ImageRpc;
+import com.php25.mediaservice.server.dto.ImgDto;
 import com.php25.mediaservice.server.service.ImageService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
