@@ -1,6 +1,7 @@
 package com.php25.userservice.server.service.impl;
 
 import com.php25.common.core.dto.DataGridPageDto;
+import com.php25.common.core.exception.ServiceException;
 import com.php25.common.core.service.IdGeneratorService;
 import com.php25.common.core.service.ModelToDtoTransferable;
 import com.php25.common.core.specification.SearchParamBuilder;
@@ -31,9 +32,9 @@ import java.util.stream.Collectors;
  * Created by penghuiping on 16/8/12.
  */
 @Slf4j
-@Transactional
 @Service
 @Primary
+@Transactional(rollbackFor = ServiceException.class)
 public class AdminRoleServiceImpl implements AdminRoleService {
 
     private AdminRoleRepository adminRoleRepository;
@@ -65,21 +66,21 @@ public class AdminRoleServiceImpl implements AdminRoleService {
     public Optional<AdminRoleDto> save(AdminRoleDto obj) {
         Assert.notNull(obj, "adminRoleDto不能为null");
         Assert.notEmpty(obj.getMenus(), "adminRoleDto.menus至少需要包含一个元素");
-        AdminRole adminRole = new AdminRole();
+        var adminRole = new AdminRole();
         BeanUtils.copyProperties(obj, adminRole);
 
         if (null == obj.getId() || "".equals(obj.getId())) {
             adminRole.setCreateTime(new Date());
         }
-        List<AdminMenuButton> adminMenuButtons = obj.getMenus().stream().map(temp -> {
+        var adminMenuButtons = obj.getMenus().stream().map(temp -> {
             AdminMenuButton adminMenuButton = new AdminMenuButton();
             BeanUtils.copyProperties(temp, adminMenuButton);
             return adminMenuButton;
         }).collect(Collectors.toList());
         adminRole.setUpdateTime(new Date());
         adminRole = adminRoleRepository.save(adminRole);
-        for (AdminMenuButton menuButton : adminMenuButtons) {
-            RoleMenu roleMenu = roleMenuRepository.findOneByRoleIdAndMenuId(adminRole.getId(), menuButton.getId());
+        for (var menuButton : adminMenuButtons) {
+            var roleMenu = roleMenuRepository.findOneByRoleIdAndMenuId(adminRole.getId(), menuButton.getId());
             if (null == roleMenu) {
                 roleMenu = new RoleMenu();
                 roleMenu.setId(idGeneratorService.getModelPrimaryKeyNumber().longValue());
@@ -88,7 +89,7 @@ public class AdminRoleServiceImpl implements AdminRoleService {
                 roleMenuRepository.save(roleMenu);
             }
         }
-        AdminRoleDto adminRoleDto = new AdminRoleDto();
+        var adminRoleDto = new AdminRoleDto();
         BeanUtils.copyProperties(adminRole, adminRoleDto);
         return Optional.of(adminRoleDto);
     }
@@ -104,15 +105,15 @@ public class AdminRoleServiceImpl implements AdminRoleService {
     public Optional<List<AdminRoleDto>> findAll(Iterable<Long> ids, Boolean lazy) {
         Assert.notEmpty((List<Long>) ids, "ids集合至少需要包含一个元素");
         Assert.notNull(lazy, "lazy不能为null");
-        List<AdminRole> adminRoles = (List<AdminRole>) adminRoleRepository.findAllById(ids);
+        var adminRoles = (List<AdminRole>) adminRoleRepository.findAllById(ids);
         return Optional.ofNullable(adminRoles.parallelStream().map(adminRole -> {
-            AdminRoleDto adminRoleDto = new AdminRoleDto();
+            var adminRoleDto = new AdminRoleDto();
             if (lazy) {
                 BeanUtils.copyProperties(adminRole, adminRoleDto, "adminMenuButtons");
             } else {
                 BeanUtils.copyProperties(adminRole, adminRoleDto);
                 List<AdminMenuButtonDto> adminMenuButtonDtos = adminRole.getAdminMenuButtons().stream().map(adminMenuButton -> {
-                    AdminMenuButtonDto adminMenuButtonDto = new AdminMenuButtonDto();
+                    var adminMenuButtonDto = new AdminMenuButtonDto();
                     BeanUtils.copyProperties(adminMenuButton, adminMenuButtonDto);
                     return adminMenuButtonDto;
                 }).collect(Collectors.toList());
@@ -124,9 +125,9 @@ public class AdminRoleServiceImpl implements AdminRoleService {
 
     @Override
     public Optional<List<AdminRoleDto>> findAllEnabled() {
-        List<AdminRole> adminRoles = adminRoleRepository.findAllEnabled();
+        var adminRoles = adminRoleRepository.findAllEnabled();
         return Optional.ofNullable(adminRoles.parallelStream().map(adminRole -> {
-            AdminRoleDto adminRoleDto = new AdminRoleDto();
+            var adminRoleDto = new AdminRoleDto();
             BeanUtils.copyProperties(adminRole, adminRoleDto, "adminMenuButtons");
             return adminRoleDto;
         }).collect(Collectors.toList()));
@@ -134,9 +135,9 @@ public class AdminRoleServiceImpl implements AdminRoleService {
 
     @Override
     public Optional<List<AdminRoleDto>> findAll() {
-        List<AdminRole> adminRoles = (List<AdminRole>) adminRoleRepository.findAll();
+        var adminRoles = (List<AdminRole>) adminRoleRepository.findAll();
         return Optional.ofNullable(adminRoles.parallelStream().map(adminRole -> {
-            AdminRoleDto adminRoleDto = new AdminRoleDto();
+            var adminRoleDto = new AdminRoleDto();
             BeanUtils.copyProperties(adminRole, adminRoleDto, "adminMenuButtons");
             return adminRoleDto;
         }).collect(Collectors.toList()));
@@ -145,7 +146,7 @@ public class AdminRoleServiceImpl implements AdminRoleService {
 
     @Override
     public Optional<DataGridPageDto<AdminRoleDto>> query(Integer pageNum, Integer pageSize, String searchParams) {
-        Optional<DataGridPageDto<AdminRoleDto>> userPage = baseService.query(pageNum, pageSize, searchParams, (adminRole, adminRoleDto) -> {
+        var userPage = baseService.query(pageNum, pageSize, searchParams, (adminRole, adminRoleDto) -> {
             BeanUtils.copyProperties(adminRole, adminRoleDto, "adminMenuButtons");
         }, Sort.Direction.DESC, "id");
         return userPage;
@@ -154,11 +155,11 @@ public class AdminRoleServiceImpl implements AdminRoleService {
     @Override
     public Optional<AdminRoleDto> findOne(Long id) {
         Assert.notNull(id, "id不能为null");
-        AdminRole adminRole = adminRoleRepository.findById(id).orElse(null);
+        var adminRole = adminRoleRepository.findById(id).orElse(null);
         if (null == adminRole) {
             return Optional.empty();
         }
-        AdminRoleDto adminRoleDto = new AdminRoleDto();
+        var adminRoleDto = new AdminRoleDto();
         BeanUtils.copyProperties(adminRole, adminRoleDto);
         return Optional.of(adminRoleDto);
     }
